Implements MixxxDB Insert, InsertMany and FindByPath

+++ mixxxdb/crates.go
@@ -2,7 +2,7 @@ package mixxxdb
 
 type Crate struct {
 	// The crate's ID in Mixxx's DB. Can be found in `crates` table
-	ID uint
+	ID int64
 	// The name of the crate.
 	Name string
 }
@@ -12,10 +12,10 @@ type CrateTracks struct {
 	Tracks []Track
 }
 
-func NewCrate(id uint, name string) Crate {
+func NewCrate(id int64, name string) Crate {
 	return Crate{ID: id, Name: name}
 }
 
-func NewCrateTracks(id uint, name string, tracks []Track) CrateTracks {
+func NewCrateTracks(id int64, name string, tracks []Track) CrateTracks {
 	return CrateTracks{Crate: Crate{ID: id, Name: name}, Tracks: tracks}
 }
